server/requirements.txt
psycopg2-binary==2.8.6
Django==3.1.3
django-redis==4.12.1
django-rest-swagger==2.2.0
djangorestframework==3.12.1
dj-database-url==0.5.0
pillow==8.0.1
gunicorn==20.0.4
django-debug-toolbar==3.2
django-rest-swagger==2.2.0

